Added configurable max history fall behind
The maximum amount of time that the connector will ever try to sync from
this historical buffer is now configurable. This will dramatically
speed up the time it takes the connector to push recent data after a
period of inactivity. The connector will simply ignore data that is not
current.